requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
redis==7.1.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
wrapt==2.0.1
